Refactor exec and resize functions
Update the ExecConfig struct to use more descriptive field names. Modify the Exec function to use the new field names. Also, update the Resize function to accept a ResizeOptions struct instead of separate height and width parameters.

Fix the Stderr field in the PodExecOptions struct to only attach stderr if tty is enabled.

Update the doResize function to use the new ResizeOptions struct when calling the Resize function.

Update the handleDirectTcpip function to use the new field names in the ExecConfig struct.

Update the import statements in the go.mod file to include new dependencies.